# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: Astra Gateway API
  version: 1.0.0
  description: |
    Enterprise AI Gateway Platform — unified access to multiple LLM providers with intelligent routing,
    governance, observability, Agent-to-Agent (A2A) communication, and Model Context Protocol (MCP) support.

    **Services & Ports (local dev)**
    | Service             | Port |
    |---------------------|------|
    | Gateway Service     | 8080 |
    | MCP Service         | 8081 |
    | A2A Service         | 8082 |
    | Auth Service        | 8083 |
    | Routing Engine      | 8084 |
    | Governance Engine   | 8085 |
    | Observability       | 8086 |
  contact:
    name: Astra Gateway Support
    url: https://docs.astragateway.io
    email: support@astragateway.io
  license:
    name: Proprietary
    url: https://astragateway.io/license
servers:
  - url: https://api.astragateway.io/v1
    description: Production (Gateway)
  - url: https://staging.astragateway.io/v1
    description: Staging (Gateway)
  - url: http://localhost:8080/v1
    description: Local — Gateway Service
  - url: http://localhost:8081/v1
    description: Local — MCP Service
  - url: http://localhost:8082/v1
    description: Local — A2A Service
  - url: http://localhost:8083/v1
    description: Local — Auth Service
  - url: http://localhost:8084/v1
    description: Local — Routing Engine
  - url: http://localhost:8085/v1
    description: Local — Governance Engine
  - url: http://localhost:8086/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /agents/register:
    post:
      tags:
        - A2A
      summary: Register an agent
      description: |
        Stores agent metadata in Redis with a 5-minute TTL (refreshed on heartbeat).
        Builds secondary indices for capability-based and region-based lookup.
      operationId: registerAgent
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistrationRequest'
            example:
              agent_id: agent-001
              name: DataAnalysisAgent
              capabilities:
                - data_analysis
                - chart_generation
              region: us-east-1
              endpoint: http://agent-001:9000
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRegistrationResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /agents/search:
    get:
      tags:
        - A2A
      summary: Search agents by capability and region
      description: |
        Queries the Redis capability index for agents that support the given capability.
        Optionally intersects with the region index to further filter results.
      operationId: searchAgents
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      parameters:
        - name: capability
          in: query
          required: true
          description: Capability to search for (e.g. `data_analysis`)
          schema:
            type: string
        - name: region
          in: query
          required: false
          description: Optional region filter (e.g. `us-east-1`)
          schema:
            type: string
      responses:
        '200':
          description: Matching agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
  /agents/{agentId}:
    get:
      tags:
        - A2A
      summary: Get agent details
      operationId: getAgent
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRecord'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - A2A
      summary: Unregister an agent
      description: |
        Removes the agent from the Redis registry and all capability/region indices.
      operationId: unregisterAgent
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent unregistered
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  status:
                    type: string
                    example: unregistered
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────
  # A2A SERVICE — Messaging
  # ─────────────────────────────────────────────
  /messages/send:
    post:
      tags:
        - A2A
      summary: Send a message to an agent
      description: |
        Pushes a message to the target agent's Redis queue (LPOP-based delivery)
        and publishes it to the `a2a.messages` Kafka topic for audit and fan-out.
        Queue TTL is 30 minutes.
      operationId: sendMessage
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            example:
              from_agent_id: agent-001
              to_agent_id: agent-002
              payload:
                task: summarise
                data: [1, 2, 3]
      responses:
        '202':
          description: Message accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageAcceptedResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /messages/publish:
    post:
      tags:
        - A2A
      summary: Publish an event to all subscribers
      description: |
        Publishes an event to the `a2a.events` Kafka topic.
        The event type is used as the Kafka message key for partitioning.
      operationId: publishEvent
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishEventRequest'
            example:
              event_type: task.completed
              agent_id: agent-001
              payload:
                result: success
                duration_ms: 1234
      responses:
        '202':
          description: Event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventPublishedResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /messages/receive:
    get:
      tags:
        - A2A
      summary: Receive pending messages for an agent
      description: |
        Pops up to 10 messages from the agent's Redis queue (RPOP) in a single call.
        Returns immediately with whatever messages are available — no long-polling.
      operationId: receiveMessages
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      parameters:
        - name: agent_id
          in: query
          required: true
          schema:
            type: string
          description: The agent ID whose queue to read
      responses:
        '200':
          description: Pending messages (may be empty)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiveMessagesResponse'

  # ─────────────────────────────────────────────
  # A2A SERVICE — Tasks
  # ─────────────────────────────────────────────
  /tasks/distribute:
    post:
      tags:
        - A2A
      summary: Distribute a task to target agents
      description: |
        Creates a task record in Redis (1-hour TTL) and fans it out to each agent
        in `target_agents` via their Redis queues. Also publishes to the `a2a.tasks` Kafka topic.
      operationId: distributeTask
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DistributeTaskRequest'
            example:
              task_type: data_analysis
              target_agents:
                - agent-002
                - agent-003
              payload:
                dataset: sales_q1.csv
      responses:
        '202':
          description: Task distributed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDistributedResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /tasks/{taskId}:
    get:
      tags:
        - A2A
      summary: Get task status
      description: Retrieves the task record from Redis, including its status and metadata.
      operationId: getTaskStatus
      servers:
        - url: http://localhost:8082/v1
          description: Local — A2A Service
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRecord'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token verified by the Auth Service (`POST /auth/verify`)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        API key validated by the Auth Service (`POST /auth/api-key/validate`).
        Format: `Bearer sk-astra-<key>`
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error summary
        message:
          type: string
          description: Detailed error message
...
    AgentRegistrationRequest:
      type: object
      required:
        - agent_id
      properties:
        agent_id:
          type: string
          example: agent-001
        name:
          type: string
        capabilities:
          type: array
          items:
            type: string
          example: [data_analysis, chart_generation]
        region:
          type: string
          example: us-east-1
        endpoint:
          type: string
          format: uri
        metadata:
          type: object
          additionalProperties: true
    AgentRegistrationResponse:
      type: object
      properties:
        agent_id:
          type: string
        status:
          type: string
          example: registered
        timestamp:
          type: integer
    AgentRecord:
      type: object
      properties:
        agent_id:
          type: string
        name:
          type: string
        capabilities:
          type: array
          items:
            type: string
        region:
          type: string
        endpoint:
          type: string
        status:
          type: string
          enum: [active, inactive]
        registered_at:
          type: integer
        last_heartbeat:
          type: integer
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AgentRecord'
        count:
          type: integer

    # ── A2A — Messaging ───────────────────────
    SendMessageRequest:
      type: object
      properties:
        from_agent_id:
          type: string
        to_agent_id:
          type: string
        payload:
          type: object
          additionalProperties: true
    MessageAcceptedResponse:
      type: object
      properties:
        message_id:
          type: string
          format: uuid
        status:
          type: string
          example: accepted
    PublishEventRequest:
      type: object
      properties:
        event_type:
          type: string
          example: task.completed
        agent_id:
          type: string
        payload:
          type: object
          additionalProperties: true
    EventPublishedResponse:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        status:
          type: string
          example: published
    ReceiveMessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            type: object
            description: Message envelope with `message_id`, `timestamp`, and original payload
        count:
          type: integer

    # ── A2A — Tasks ───────────────────────────
    DistributeTaskRequest:
      type: object
      properties:
        task_type:
          type: string
          example: data_analysis
        target_agents:
          type: array
          items:
            type: string
          description: Agent IDs to receive this task; omit to publish only to Kafka
        payload:
          type: object
          additionalProperties: true
    TaskDistributedResponse:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          example: distributed
    TaskRecord:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        task_type:
          type: string
        status:
          type: string
          example: distributed
        target_agents:
          type: array
          items:
            type: string
        created_at:
          type: integer
        payload:
          type: object

    # ── MCP ───────────────────────────────────
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'